tests/validation/requirements.txt
pytest==3.8.1
pytest-repeat==0.7.0
pytest-xdist==1.23.0
boto3==1.10.6
botocore==1.13.6
cryptography==3.4.7
flake8==3.5.0
invoke==1.2.0
Jinja2==3.0.0
kubernetes==12.0.1
paramiko==2.10.1
python-digitalocean==1.13.2
PyYAML==5.4.1
packaging==20.9
